requirements.txt
azure-cosmos
cli_validator
numpy
openai
